requirements.txt
anaconda-anon-usage @ file:///C:/miniconda3/conda-bld/anaconda-anon-usage_1757625796025/work
anaconda-auth @ file:///C:/miniconda3/conda-bld/anaconda-cloud-auth-split_1759943207790/work
anaconda-cli-base @ file:///C:/miniconda3/conda-bld/anaconda-cli-base_1760698255040/work
annotated-types @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/annotated-types_1729043709318/work
archspec @ file:///home/task_175812491784513/conda-bld/archspec_1758124989039/work
attrs==25.4.0
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi==2025.11.12
cffi @ file:///C:/miniconda3/conda-bld/cffi_1760098048206/work
charset-normalizer==3.4.4
click @ file:///C:/b/abs_42rbjngfys/croot/click_1744271606983/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
conda @ file:///C:/miniconda3/conda-bld/conda_1760457894461/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1755883788794/work
conda-anaconda-tos @ file:///C:/b/abs_c9yejtsx9g/croot/conda-anaconda-tos_1755123332641/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_2d41dz5u60/croot/conda-package-streaming_1751307521903/work
cryptography @ file:///C:/miniconda3/conda-bld/cryptography-split_1760106330961/work
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
frozendict @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/frozendict_1729090952168/work
h11==0.16.0
idna==3.11
iniconfig==2.3.0
jaraco.classes @ file:///C:/b/abs_6erueoob1v/croot/jaraco.classes_1755516340851/work
jaraco.context @ file:///C:/Users/dev-admin/buildout/perseverance-python-buildout/croot/jaraco.context_1731721000658/work
jaraco.functools @ file:///C:/b/abs_a1jv6v_pzp/croot/jaraco.functools_1740408129620/work
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer @ file:///C:/b/abs_73u73l7pl9/croot/jsonpointer_1753788460913/work
keyring @ file:///C:/b/abs_596d81ux2c/croot/keyring_1745484709696/work
libmambapy @ file:///C:/miniconda3/conda-bld/mamba-split_1757623139479/work/libmambapy
markdown-it-py @ file:///C:/b/abs_3espcrdl35/croot/markdown-it-py_1756299213703/work
mdurl @ file:///C:/miniconda3/conda-bld/mdurl_1758552277768/work
menuinst @ file:///C:/b/abs_1bepgi4wpa/croot/menuinst_1753464604355/work
more-itertools @ file:///C:/miniconda3/conda-bld/more-itertools_1761121564395/work
outcome==1.3.0.post0
packaging @ file:///C:/miniconda3/conda-bld/packaging_1761049099114/work
pkce @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pkce_1729049216383/work
platformdirs @ file:///C:/b/abs_ddh15014or/croot/platformdirs_1744273060660/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496153123/work
pydantic @ file:///C:/miniconda3/conda-bld/pydantic_1760697509246/work
pydantic-settings @ file:///C:/miniconda3/conda-bld/pydantic-settings_1758784270277/work
pydantic_core @ file:///C:/miniconda3/conda-bld/pydantic-core_1760597457096/work
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
PyJWT @ file:///C:/Users/dev-admin/buildout/croot/pyjwt_1739559988030/work
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
pytest==9.0.1
python-dotenv==1.2.1
pywin32-ctypes @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pywin32-ctypes_1729046491215/work
readchar @ file:///C:/miniconda3/conda-bld/readchar_1760613474723/work
requests @ file:///C:/b/abs_f53smfvswb/croot/requests_1756709376206/work
rich @ file:///C:/miniconda3/conda-bld/rich_1760375661587/work
ruamel.yaml @ file:///C:/b/abs_0dn1nhdq75/croot/ruamel.yaml_1745960327187/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
selenium==4.38.0
semver @ file:///C:/b/abs_6bgol0at6h/croot/semver_1736542515243/work
setuptools==80.9.0
shellingham @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/shellingham_1729046672855/work
sniffio==1.3.1
sortedcontainers==2.4.0
tomli @ file:///C:/b/abs_88e598m6o8/croot/tomli_1753774604115/work
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
trio==0.32.0
trio-websocket==0.12.2
truststore @ file:///C:/b/abs_edelxif8va/croot/truststore_1753907192099/work
typer @ file:///C:/miniconda3/conda-bld/typer_1757580777256/work
typing-inspection @ file:///C:/miniconda3/conda-bld/typing-inspection_1760614188477/work
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
webdriver-manager==4.0.2
websocket-client==1.9.0
wheel==0.45.1
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
wsproto==1.3.2
zstandard @ file:///C:/miniconda3/conda-bld/zstandard_1758189089298/work
